fix: x,z report printing

+++ BDO_DatecsDP25_2.0/Dp25.cs
@@ -401,7 +401,7 @@ namespace BDO_DatecsDP25
         public PrintReportResponse PrintReport(Commands.ReportType type)
         {
             var Command = 69;
-            var Data = ((int)type) + "\t";
+            var Data = (type) + "\t";
             return new PrintReportResponse(SendMessage(Command, Data));
         }
